fix dtype coming out of GrayscaleEnv
This is especially relevant for game environments where the observation is a uint8 tensor.

+++ anyrl/envs/wrappers/image.py
@@ -56,7 +56,7 @@ class GrayscaleEnv(gym.ObservationWrapper):
             observation = observation // 3
         else:
             observation = observation / 3
-        return np.sum(observation, axis=-1, keepdims=self._keep_depth)
+        return np.sum(observation, axis=-1, keepdims=self._keep_depth, dtype=observation.dtype)
 
 class FrameStackEnv(gym.Wrapper):
     """
